-- many fixes for win32 port
git-svn-id: https://svn.rtmpd.com/crtmpserver/trunk@49 01689507-d990-df11-9d7d-0017a411d416

+++ sources/applications/applestreamingclient/include/protocols/variant/messagestructure.h
@@ -129,7 +129,7 @@ do { \
 do {\
 	ASC_RES_DEBUG_FILE(r); \
 	ASC_RES_DEBUG_LINE_NUMBER(r); \
-	ASC_RES_STATUS(r)=(status); \
+	ASC_RES_STATUS(r)=(uint32_t) (status); \
 	switch ((status)) { \
 		case ASC_RES_STATUS_OK: \
 		{ \
