+++ sources/thelib/src/mediaformats/flv/flvdocument.cpp
@@ -156,7 +156,7 @@ bool FLVDocument::BuildFrames() {
 		//12. Read the metadata or ignore the data payload
 		if (frame.type == MEDIAFRAME_TYPE_DATA) {
 			IOBuffer tempBuffer;
-			tempBuffer.ReadFromFs(_mediaFile, frame.length);
+			tempBuffer.ReadFromFs(_mediaFile, (uint32_t)frame.length);
 			//tempBuffer.PutInputBuffer(&buffer, 0, length);
 #ifdef HAS_PROTOCOL_RTMP
 			AMF0Serializer amfSerializer;
