+++ sources/thelib/src/protocols/rtmp/streaming/infilertmpflvstream.cpp
@@ -39,7 +39,7 @@ bool InFileRTMPFLVStream::BuildFrame(FileClass *pFile, MediaFrame &mediaFrame,
     }
 
     //2. Read the data
-    if (!buffer.ReadFromFs(*pFile, mediaFrame.length)) {
+    if (!buffer.ReadFromFs(*pFile, (uint32_t)mediaFrame.length)) {
         FATAL("Unable to read %llu bytes from offset %llu", mediaFrame.length, mediaFrame.start);
         return false;
     }
